[fix] optimized + separated light sources

+++ src/geometry.rs
@@ -2,10 +2,10 @@ use arrayvec::ArrayVec;
 use na::Matrix3;
 use nalgebra::{UnitQuaternion, Vector3};
 
-pub type Fp = f64;
-pub const FP_PI: Fp = std::f64::consts::PI;
-pub const FP_INF: Fp = f64::INFINITY;
-pub const FP_NEG_INF: Fp = f64::NEG_INFINITY;
+pub type Fp = f32;
+pub const FP_PI: Fp = std::f32::consts::PI;
+pub const FP_INF: Fp = f32::INFINITY;
+pub const FP_NEG_INF: Fp = f32::NEG_INFINITY;
 pub type Vec3f = Vector3<Fp>;
 
 #[derive(Clone, Debug)]
